requirements.txt
flask==3.0.3
requests==2.32.3
google-generativeai==0.8.2
pillow==10.4.0